Fix bug in FolderImageLoader
The FolderImageLoader chooses a list of possible integers
among all integers of indices of the images in the folder
that we load.

In FolderImageLoader we choose the indices of the images
to sample every time, from all possible indices of the images in
the folder.
The selection was between 0 to num_of_images-1, so when the folder
contains only one image, we tried to choose indices from
an empty group.
The indices now are selected from 0 to num_of_images, so the bug
is fixed.