Update releases scripts
This commit changes the release script to use the PyPi server for publishing
the package. Also, this commit adds a script to test the release on testpypi.
The test script generates a tag, pushes it, uploads the package to testpypi
and removes the tag (as the actual release will need this tag).