Add different hardware model support (#118)
Add a way to model the hardware-related settings to consider during optimization.

Move parts from QuantizationConfig that contained information about the intended hardware the optimized model will use for inference to a new class called HardwareModel. That includes the definition and configuration of operators and the definition of fusing patterns.

A new class called FrameworkHardwareModel was created as well. This class attaches layers’ representation of a framework to the operators model for a defined HardwareModel.

A new function called get_model was also added to ease the use of users to get ready models for Tensorflow and Pytorch.

Documentation and tutorials were updated correspondingly.

Co-authored-by: Reuven Peretz <reuvenp@altair-semi.com>