Change GPTQ node search method
Add mapping in Keras model builder from layers to their graph nodes.
This is since there are times we rely on the layer's name (during cloning in GPTQ, MP, etc.)
but the layer's in the final model can be different from the node's name,
so we save a mapping of the layer's output tensor to the node.