Fix inferable quantizers serialization and inference issues
Fix issues with inferable quantizers:
1. Keras quantizers should receive list in their init instead of np arrays. The
reason for that is that during deserialization keras build the quantizers with lists
(even if they were saved as np arrays).
2. In LUT weight quantizers the kernel had dtype of float64 which failed inference.
The reason for this is that by default numpy arrays are float64. This commit fixes this
by explicity stating the dtype to be float32.