Fix tf export of reused layers
In TF export, as part of weights building of a layer, the attribute input_shape
of the layer was searched. When a layer has multiple inbound layers with different shapes
this is impossible. For this reason, we get the first input shape of the layer to support
cases where multiple input shapes exist.