Fix tf export of reused layers (#684)
In TF export, as part of the weights building of a layer, the attribute input_shape
of the layer was searched. This is impossible when a layer has multiple inbound layers with different shapes. For this reason, we get the first input shape of the layer to support
cases where multiple input shapes exist.


---------

Co-authored-by: reuvenp <reuvenp@altair-semi.com>