Pytorch export fixes (#691)
A _substitute_fully_quantized_model method was added to base pytorch exporter, where a substitution is performed for "fully-quantized" PyTorch models. The method uses the weight quantizers from PytorchQuantizationWrapper layers to quantize the weights and set them in the layer. Then, it replaces all wrapped layers with the layers their internal layers.

The torchscript exported models have two changes applied: 1) Wrapped layers are replaced with internal layers. 2) Weights are quantized and set as layer weights (instead of quantizing them on-the-fly).

ONNX exported models already had the quantized weights set before export (thus, they were not quantized on-the-fly). ONNX exporter change is that wrapped layers are replaced with internal layers.

---------

Co-authored-by: reuvenp <reuvenp@altair-semi.com>