Fix params search in pytorch SNC
In SNC, when better shifting values are tested for better choice,
the histogram (which is a numpy object) is quantized using the non-linear node activation
quantization function (to estimate the expected mse comparing to the original histogram).
The issue is that this quantization function is a framewrok tensor, which makes it fail since it
expoects a fw tensor. To overcome the issue, this commit adds to the commmon part of SNC an argument whichis a callable
that receives two argument and returns one: it gets the fw activation quantization function
and the bins to quantize. The function (of each fw) responsible for doing (if needed) a preprocessing and postprocessing
to the bins which is a numpy object.
Only if this function is passed - common SNC will use this function. If not - it assumes
no processing is needed and simply uses the activation quantization quantizer to quantize the bins (like tf for example).