Data Generation (#760)
This pull request adds the Data Generation Library to the Model Compression Toolkit (MCT). It is a powerful Python package designed to facilitate the generation of synthetic data. This library leverages the statistics stored in the model's batch normalization (BatchNorm) layers to create synthetic data that faithfully represent the model's training data characteristics. This generated data is valuable for various compression tasks where representative data of the model's training set is needed, such as quantization and pruning.