Create Softmax layer in activation decomposition instead of Activation layer
In MP we need to know the axis used in Softmax for KL-Divergence computation.
In activation decomposition, we created a simple 'Activation' layer thus we can not
use the axis info when computing MP distance. This commit changes the decomposition
to create a Softmax layer, thus enabling the KL-Divergence in MP distance computation.