Add a flag to indicate whether to run mixed precision or single precision according to resource utilization provided by user. (#1052)
Implements a function to select precision mode for model quantization.

This PR adds a new function that evaluates whether model quantization should proceed using mixed precision or can be effectively performed using single precision. This decision is based on the resource utilization constraints specified by the user. The addition aims to optimize the quantization process by adapting it to the specific resource capacities and performance targets of the deployment environment.